feat(save-editor): Charaktere tab - master list + 4 detail sub-tabs
Replace the separate Attribute + Inventory top-level tabs with one
Charaktere tab: a shared CharacterMasterList on the left and four detail
sub-tabs (Attribute, Inventar, Wissen, Ereignisse) on the right, all
driven by the shared state.selectedActor selection.

- Extract the Attribute detail body into attribute_detail.dart
  (AttributeDetail) and the Inventory detail body into
  inventory_detail.dart (InventoryDetail), parameterized by `actor`;
  moved their exclusive private widgets/helpers verbatim.
- Add characters_tab.dart (CharactersTab) hosting the master list + a
  DefaultTabController(4) whose sub-tab bodies are kept alive; orphan
  selections get a clean empty state for Attribute/Inventar/Ereignisse
  and only Wissen is wired.
- Swap the top-level bar to length 6 (Overview, Charaktere, Progression,
  All data, Backups, Settings); delete the dead _AttributePanel /
  _InventoryPanel shells.
- Add l10n keys tabCharacters + characterNoActorBody across all locales
  and regenerate app_localizations.
- Retarget widget_test + the six npc_* tests at the new structure
  (navigate via the Characters tab; fakes answer private.characters.list).
