+++ apps/save-editor/lib/features/editor/ui/editor_page.dart
@@ -12,13 +12,8 @@ import 'package:goresave/features/app/ui/window_chrome.dart';
 import 'package:goresave/features/editor/domain/editor_notifier.dart';
 import 'package:goresave/features/editor/domain/editor_models.dart';
 import 'package:goresave/features/editor/domain/game_time.dart';
-import 'package:goresave/features/editor/domain/item_categories.dart';
-import 'package:goresave/features/editor/domain/npc_attributes.dart';
-import 'package:goresave/features/editor/ui/actor_detail_header.dart';
-import 'package:goresave/features/editor/ui/actor_selector.dart';
-import 'package:goresave/features/editor/ui/npc_attributes_panel.dart';
-import 'package:goresave/features/editor/ui/sidebar_tile.dart';
 import 'package:goresave/features/editor/domain/pending_edits.dart';
+import 'package:goresave/features/editor/ui/characters_tab.dart';
 import 'package:goresave/features/localization/domain/localization_controller.dart';
 import 'package:goresave/features/localization/ui/localization_flow.dart';
 import 'package:goresave/features/localization/ui/localization_settings.dart';
@@ -26,10 +21,7 @@ import 'package:goresave/l10n/app_localizations.dart';
 import 'package:goresave/loc/loc_catalog_provider.dart';
 import 'package:goresave/providers/data_providers.dart';
 import 'package:intl/intl.dart';
-import 'add_inventory_item_dialog.dart';
 import 'difficulty_dialog.dart';
-import '../domain/hero_attributes.dart' show TypedValueEdit;
-import 'hero_stats_card.dart';
 import 'progression_panel.dart';
 
 final _bytes = NumberFormat.decimalPattern();
@@ -668,7 +660,7 @@ class _EditorWorkspace extends StatelessWidget {
       final inspection = state.inspection!;
       final pendingCount = state.pendingEditCount;
       content = DefaultTabController(
-        length: 7,
+        length: 6,
         child: Column(
           children: [
             Container(
@@ -684,12 +676,8 @@ class _EditorWorkspace extends StatelessWidget {
                           text: l10n.tabOverview,
                         ),
                         Tab(
-                          icon: const Icon(Icons.person_outline),
-                          text: l10n.tabAttribute,
-                        ),
-                        Tab(
-                          icon: const Icon(Icons.inventory_2_outlined),
-                          text: l10n.tabInventory,
+                          icon: const Icon(Icons.people_outline),
+                          text: l10n.tabCharacters,
                         ),
                         Tab(
                           icon: const Icon(Icons.flag_outlined),
@@ -771,21 +759,23 @@ class _EditorWorkspace extends StatelessWidget {
                     ),
                   ),
                   _KeepAliveTab(
-                    child: _AttributePanel(
+                    child: CharactersTab(
                       inspection: inspection,
                       notifier: notifier,
                       // Private writes recompress the payload, so also require the
-                      // codec to be compress-ready, not just decode-ready.
-                      editable:
+                      // codec to be compress-ready, not just decode-ready — same
+                      // gating the old Attribute tab used.
+                      attributeEditable:
                           inspection.privateEditable &&
                           state.codecCompressReady,
-                    ),
-                  ),
-                  _KeepAliveTab(
-                    child: _InventoryPanel(
-                      inspection: inspection,
-                      notifier: notifier,
-                      canCompress: state.codecCompressReady,
+                      // Same value the old Inventory tab received for canCompress.
+                      inventoryCanCompress: state.codecCompressReady,
+                      // Same gating the Progression tab uses for its knowledge/
+                      // events detail panels.
+                      progressionEditable:
+                          inspection.privateEditable &&
+                          inspection.privateTypedVerified &&
+                          state.codecCompressReady,
                     ),
                   ),
                   _KeepAliveTab(
@@ -1637,343 +1627,6 @@ class _MetricGrid extends StatelessWidget {
   }
 }
 
-/// Reverse a stored per-NPC attribute registry entry back into the panel's
-/// [NpcTypedEdit] drafts so [NpcAttributesPanel] can resume from them on a
-/// revisit. Inverse of the `private.typed.setValue` JSON the onPendingChanged
-/// handler writes. Tolerant of unexpected shapes (skips what it can't parse).
-List<NpcTypedEdit> _npcAttributeDraftsFromPending(PendingSaveEdit? pending) {
-  if (pending == null) return const [];
-  final drafts = <NpcTypedEdit>[];
-  for (final edit in pending.edits) {
-    if (edit['path'] != 'private.typed.setValue') continue;
-    final value = edit['value'];
-    if (value is! Map) continue;
-    final path = value['path'];
-    final raw = value['value'];
-    if (path is! List) continue;
-    final segments = path.whereType<String>().toList();
-    if (segments.length != path.length) continue;
-    final number = raw is num ? raw.toDouble() : null;
-    if (number == null) continue;
-    drafts.add(NpcTypedEdit(path: segments, value: number));
-  }
-  return drafts;
-}
-
-/// Reconstruct the player's queued attribute drafts from the 'heroStats' pending
-/// entry so [HeroStatsCard] can resume from them when the user returns to the
-/// Player after selecting an NPC. Mirror of [_npcAttributeDraftsFromPending]
-/// for the player's [TypedValueEdit] type (same `private.typed.setValue` JSON).
-List<TypedValueEdit> _heroDraftsFromPending(PendingSaveEdit? pending) {
-  if (pending == null) return const [];
-  final drafts = <TypedValueEdit>[];
-  for (final edit in pending.edits) {
-    if (edit['path'] != 'private.typed.setValue') continue;
-    final value = edit['value'];
-    if (value is! Map) continue;
-    final path = value['path'];
-    final raw = value['value'];
-    if (path is! List) continue;
-    final segments = path.whereType<String>().toList();
-    if (segments.length != path.length) continue;
-    final number = raw is num ? raw.toDouble() : null;
-    if (number == null) continue;
-    drafts.add(TypedValueEdit(path: segments, value: number));
-  }
-  return drafts;
-}
-
-/// The "Attribute" tab body: a shared [ActorSelector] sidebar on the left and a
-/// detail editor on the right. When the Player is selected the existing player
-/// attribute view ([_PrivatePanel]) is shown unchanged; when an NPC is selected
-/// its attributes are loaded and rendered in a flat [NpcAttributesPanel]. The
-/// selection is the SHARED actor state (read via `state.selectedActor`, written
-/// via `notifier.selectActor`) so it stays in sync with the Inventory tab.
-class _AttributePanel extends ConsumerWidget {
-  const _AttributePanel({
-    required this.inspection,
-    required this.notifier,
-    required this.editable,
-  });
-
-  final SaveInspection inspection;
-  final EditorNotifier notifier;
-  final bool editable;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context);
-    final state = ref.watch(editorProvider);
-    final selected = state.selectedActor;
-    final lang = ref.watch(currentGameLangProvider);
-    final locCatalog =
-        ref.watch(locCatalogProvider).asData?.value ?? const {};
-
-    final Widget detail;
-    if (selected.isPlayer) {
-      // Player → a shared header ("Player", no GlobalId) above the EXISTING
-      // player attribute view, unchanged below.
-      detail = Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          ActorDetailHeader(
-            actor: selected,
-            locCatalog: locCatalog,
-            lang: lang,
-          ),
-          Expanded(
-            child: _PrivatePanel(
-              icon: Icons.person_outline,
-              title: l10n.tabPlayer,
-              isPlayer: true,
-              inspection: inspection,
-              notifier: notifier,
-              editable: editable,
-              lockedBody: l10n.playerLockedBody,
-            ),
-          ),
-        ],
-      );
-    } else {
-      // NPC → flat attribute editor wired to a PER-NPC pending-edit key so it
-      // never collides with the player's 'heroStats' contribution AND each NPC's
-      // edits accumulate independently. A shared 'npc.attributes' key would let
-      // a switch from NPC-A to NPC-B leave A's edit lingering under the same key
-      // (the panel reload only clears its local field state, not the registry),
-      // so A's edit would silently apply on the next Save while the UI shows B.
-      // Keying by the NPC's GlobalId means A stays under 'npc.attributes:A' and
-      // B under 'npc.attributes:B'; the save flow batches every key and the
-      // distinct per-NPC GlobalId typed paths never conflict.
-      final npcId = selected.id!;
-      final pendingKey = 'npc.attributes:$npcId';
-      // Status row: `Status: <lebend|tot>` + a Revive action, rendered as the
-      // FIRST entry of the core ("Hauptwerte") group detail (NPC-only). Revive
-      // drives a STANDALONE structural edit the core won't batch with peers, so
-      // it REGISTERS a per-NPC pending edit (`npc.revive:$id`) — the global Save
-      // button applies it (saveAllPending splits each splicing edit into its own
-      // write_save). No file write happens on tap. The pending flag is read from
-      // the registry so a queued revive reflects optimistically before Save; the
-      // reloadKey matches the panel's so a Save's trailing refresh re-seeds the
-      // status line/HP from disk.
-      final revivePending = state.pendingEdits.containsKey('npc.revive:$npcId');
-      final statusConfig = NpcStatusConfig(
-        npcId: npcId,
-        editable: editable,
-        // Carry the dead state known from the sidebar selection as a fallback
-        // for when the async summary reload fails / the id is missing.
-        knownDead: selected.isDead,
-        reloadKey: (inspection, npcId),
-        // Use the cached FULL NPC list (shared with the selector, one decompress)
-        // and let the row pick the exact-id match. A substring query with the
-        // default 100-row page could miss this NPC when many ids match, leaving
-        // the action disabled and isDead wrongly defaulting.
-        load: () => notifier.loadAllNpcActors(),
-        onRevive: () => notifier.setPendingNpcRevive(npcId),
-        revivePending: revivePending,
-      );
-      detail = Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          ActorDetailHeader(
-            actor: selected,
-            locCatalog: locCatalog,
-            lang: lang,
-          ),
-          Expanded(
-            child: NpcAttributesPanel(
-              // Reload when the inspected save OR the selected NPC changes, so
-              // the list re-seeds from this NPC's saved values.
-              reloadKey: (inspection, npcId),
-              load: () => notifier.loadNpcAttributes(npcId),
-              editable: editable,
-              // Status row lives at the top of the core ("Hauptwerte") group.
-              status: statusConfig,
-              // Resume from this NPC's queued attribute drafts on revisit:
-              // reverse the stored typed.setValue edits back into the panel's
-              // NpcTypedEdit drafts. Without this, returning to a previously
-              // edited NPC and editing another attribute would replace the
-              // stored entry with only the newly-dirty field, dropping the rest.
-              initialPending: () => _npcAttributeDraftsFromPending(
-                notifier.pendingEditFor(pendingKey),
-              ),
-              onPendingChanged: (edits, validationError) {
-                if (validationError != null) {
-                  // Transient invalid/empty input in one field must NOT discard
-                  // the NPC's already-stored valid drafts (switching actors
-                  // disposes this panel and loses its local _pending). Keep the
-                  // stored drafts but BLOCK global Save while invalid, so the
-                  // now-stale stored value is never written behind the bad field.
-                  notifier.setNpcEditInvalid(pendingKey);
-                  return;
-                }
-                notifier.setNpcEditInvalid(null);
-                if (edits.isEmpty) {
-                  notifier.clearPendingEdit(pendingKey);
-                } else {
-                  notifier.setPendingEdit(
-                    pendingKey,
-                    PendingSaveEdit(
-                      edits: [
-                        for (final edit in edits)
-                          {
-                            'path': 'private.typed.setValue',
-                            'value': {'path': edit.path, 'value': edit.value},
-                          },
-                      ],
-                    ),
-                  );
-                }
-              },
-            ),
-          ),
-        ],
-      );
-    }
-
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        SizedBox(
-          width: 280,
-          child: ActorSelector(
-            selected: selected,
-            onSelect: notifier.selectActor,
-            loadNpcs: notifier.loadAllNpcActors,
-            // Reload the NPC list when the inspected save changes (new
-            // inspection identity from a save/refresh), so the selector never
-            // shows the previous file's NPCs/badges against the new save.
-            reloadKey: inspection,
-            locCatalog: locCatalog,
-            lang: lang,
-          ),
-        ),
-        const VerticalDivider(width: 1),
-        Expanded(child: detail),
-      ],
-    );
-  }
-}
-
-class _PrivatePanel extends StatelessWidget {
-  const _PrivatePanel({
-    required this.icon,
-    required this.title,
-    required this.isPlayer,
-    required this.inspection,
-    required this.notifier,
-    required this.editable,
-    required this.lockedBody,
-  });
-
-  final IconData icon;
-  final String title;
-  final bool isPlayer;
-  final SaveInspection inspection;
-  final EditorNotifier notifier;
-  final bool editable;
-  final String lockedBody;
-
-  Widget _legacyAttributesCard() {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-        child: _PrivatePlayerAttributesEditor(
-          player: inspection.privatePlayer,
-          notifier: notifier,
-          editable: editable,
-          reloadKey: inspection,
-        ),
-      ),
-    );
-  }
-
-  Widget? _transformCard() {
-    if (inspection.privatePlayer.transform == null) return null;
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-        child: _PrivatePlayerTransformEditor(
-          transform: inspection.privatePlayer.transform!,
-          editable:
-              editable &&
-              inspection.privatePlayer.writable.contains(
-                'private.player.setTransform',
-              ),
-          notifier: notifier,
-          reloadKey: inspection,
-        ),
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (inspection.privateDecoded) {
-      // Typed path: HeroStatsCard manages its own internal scroll for the
-      // detail area and pins the sidebar. Give it the full pane via Padding
-      // (not ListView) so it has a finite height to work with.
-      if (isPlayer && inspection.privateTypedVerified) {
-        return Padding(
-          padding: const EdgeInsets.all(20),
-          child: HeroStatsCard(
-            // New SaveInspection instance after every write/refresh —
-            // changing identity drops pending edits and reloads.
-            reloadKey: inspection,
-            load: notifier.loadHeroAttributes,
-            // Resume from any unsaved player drafts after returning from an NPC.
-            initialPending: () =>
-                _heroDraftsFromPending(notifier.pendingEditFor('heroStats')),
-            onPendingChanged: (edits, validationError) {
-              if (edits.isEmpty || validationError != null) {
-                notifier.clearPendingEdit('heroStats');
-              } else {
-                notifier.setPendingEdit(
-                  'heroStats',
-                  PendingSaveEdit(
-                    edits: [
-                      for (final edit in edits)
-                        {
-                          'path': 'private.typed.setValue',
-                          'value': {'path': edit.path, 'value': edit.value},
-                        },
-                    ],
-                  ),
-                );
-              }
-            },
-            editable: editable,
-            // Spec: if the typed search errors out or finds nothing on a
-            // typed-OK save, the heuristic editor stays available.
-            fallback: inspection.privatePlayer.attributes.isNotEmpty
-                ? _legacyAttributesCard()
-                : null,
-            transformCard: _transformCard(),
-          ),
-        );
-      }
-      // Legacy / non-typed path: stacked layout in a ListView.
-      return ListView(
-        padding: const EdgeInsets.all(20),
-        children: [
-          if (isPlayer) ...[
-            // Typed parse failed or not verified: stacked legacy layout —
-            // no sidebar, no typed load call.
-            if (inspection.privatePlayer.attributes.isNotEmpty) ...[
-              _legacyAttributesCard(),
-              const SizedBox(height: 16),
-            ],
-            if (inspection.privatePlayer.transform != null) ...[
-              _transformCard()!,
-              const SizedBox(height: 16),
-            ],
-          ],
-        ],
-      );
-    }
-    return _MessagePane(icon: icon, title: title, body: lockedBody);
-  }
-}
-
 class _CollapsibleCardHeader extends StatelessWidget {
   const _CollapsibleCardHeader({
     required this.icon,
@@ -2019,1683 +1672,6 @@ class _CollapsibleCardHeader extends StatelessWidget {
       ),
     );
   }
-}
-
-/// Resolves an inventory [PrivateInventorySummary] into the editable /
-/// addable / removable gates the card consumes. [typedVerified] and
-/// [canCompress] are save-wide; [writable] is the inventory's advertised op
-/// list (the player's from the inspection, an NPC's from its loaded summary).
-({bool editable, bool canAddItem, bool canRemoveItem}) _inventoryGates({
-  required List<String> writable,
-  required bool privateEditable,
-  required bool typedVerified,
-  required bool canCompress,
-}) {
-  final editable =
-      privateEditable &&
-      canCompress &&
-      writable.contains('private.inventory.setItemCount');
-  final canAddItem =
-      privateEditable &&
-      canCompress &&
-      typedVerified &&
-      writable.contains('private.inventory.addItem');
-  final canRemoveItem =
-      privateEditable &&
-      canCompress &&
-      typedVerified &&
-      writable.contains('private.inventory.removeItem');
-  return (
-    editable: editable,
-    canAddItem: canAddItem,
-    canRemoveItem: canRemoveItem,
-  );
-}
-
-/// The "Inventory" tab body: the SAME shared [ActorSelector] sidebar as the
-/// Attribute tab on the left and the inventory detail on the right. When the
-/// Player is selected the existing inventory card path is shown UNCHANGED
-/// (loaded from the inspection, edits with no actorId). When an NPC is selected
-/// its inventory is loaded via `private.npc.inventory` and fed to the SAME card
-/// widget, with edits keyed PER-NPC and carrying the NPC's actorId. Selection
-/// is the SHARED actor state (read `state.selectedActor`, written
-/// `notifier.selectActor`) so it stays in sync with the Attribute tab.
-class _InventoryPanel extends ConsumerWidget {
-  const _InventoryPanel({
-    required this.inspection,
-    required this.notifier,
-    this.canCompress = false,
-  });
-
-  final SaveInspection inspection;
-  final EditorNotifier notifier;
-  final bool canCompress;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context);
-    if (!inspection.privateDecoded) {
-      return _MessagePane(
-        icon: Icons.inventory_2_outlined,
-        title: l10n.inventoryTitle,
-        body: l10n.inventoryNeedsDecoded,
-      );
-    }
-
-    final state = ref.watch(editorProvider);
-    final selected = state.selectedActor;
-    final lang = ref.watch(currentGameLangProvider);
-    final locCatalog =
-        ref.watch(locCatalogProvider).asData?.value ?? const {};
-
-    final Widget body;
-    if (selected.isPlayer) {
-      body = _playerInventoryDetail(context);
-    } else {
-      // NPC → load its inventory and feed the SAME card, keyed per-NPC so
-      // switching NPCs never clobbers either one's queued edits.
-      final npcId = selected.id!;
-      body = _NpcInventoryDetail(
-        // Reload when the inspected save OR the selected NPC changes.
-        key: ValueKey(('npc-inventory', npcId)),
-        reloadKey: (inspection, npcId),
-        npcId: npcId,
-        notifier: notifier,
-        privateEditable: inspection.privateEditable,
-        typedVerified: inspection.privateTypedVerified,
-        canCompress: canCompress,
-      );
-    }
-
-    // Shared header above the inventory body, identifying the selected actor
-    // (NPC name + full GlobalId, or "Player") — same widget the Attribute tab
-    // uses, so both tabs make the selection obvious.
-    final detail = Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        ActorDetailHeader(
-          actor: selected,
-          locCatalog: locCatalog,
-          lang: lang,
-        ),
-        Expanded(child: body),
-      ],
-    );
-
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        SizedBox(
-          width: 280,
-          child: ActorSelector(
-            selected: selected,
-            onSelect: notifier.selectActor,
-            loadNpcs: notifier.loadAllNpcActors,
-            // Reload the NPC list when the inspected save changes, mirroring the
-            // Attribute tab — keeps the selector in sync after a save/refresh.
-            reloadKey: inspection,
-            locCatalog: locCatalog,
-            lang: lang,
-          ),
-        ),
-        const VerticalDivider(width: 1),
-        Expanded(child: detail),
-      ],
-    );
-  }
-
-  /// The existing player inventory detail — load from the inspection, edits
-  /// under the literal `'inventory'` key with no actorId. Behaviour is
-  /// byte-for-byte unchanged from before the shared-selector refactor.
-  Widget _playerInventoryDetail(BuildContext context) {
-    final l10n = AppLocalizations.of(context);
-    final gates = _inventoryGates(
-      writable: inspection.privateInventory.writable,
-      privateEditable: inspection.privateEditable,
-      typedVerified: inspection.privateTypedVerified,
-      canCompress: canCompress,
-    );
-    final hasItems = inspection.privateInventory.hasData;
-    if (!hasItems && !gates.canAddItem && !gates.canRemoveItem) {
-      return _MessagePane(
-        icon: Icons.inventory_2_outlined,
-        title: l10n.inventoryTitle,
-        body: l10n.inventoryNoStacks,
-      );
-    }
-    return Padding(
-      padding: const EdgeInsets.all(20),
-      child: _PrivateInventorySummaryCard(
-        inventory: inspection.privateInventory,
-        notifier: notifier,
-        editable: gates.editable,
-        canAddItem: gates.canAddItem,
-        canRemoveItem: gates.canRemoveItem,
-      ),
-    );
-  }
-}
-
-/// Loads the selected NPC's inventory via `private.npc.inventory` and renders
-/// it with the SAME [_PrivateInventorySummaryCard] the player uses. Edits are
-/// keyed PER-NPC (`'inventory:$npcId'`) and stamped with the NPC's actorId.
-class _NpcInventoryDetail extends StatefulWidget {
-  const _NpcInventoryDetail({
-    super.key,
-    required this.reloadKey,
-    required this.npcId,
-    required this.notifier,
-    required this.privateEditable,
-    required this.typedVerified,
-    required this.canCompress,
-  });
-
-  final Object reloadKey;
-  final String npcId;
-  final EditorNotifier notifier;
-  final bool privateEditable;
-  final bool typedVerified;
-  final bool canCompress;
-
-  @override
-  State<_NpcInventoryDetail> createState() => _NpcInventoryDetailState();
-}
-
-class _NpcInventoryDetailState extends State<_NpcInventoryDetail> {
-  NpcInventoryResult? _result;
-  bool _loading = false;
-  int _reloadEpoch = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    _reload();
-  }
-
-  @override
-  void didUpdateWidget(covariant _NpcInventoryDetail oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (widget.reloadKey != oldWidget.reloadKey) _reload();
-  }
-
-  Future<void> _reload() async {
-    final epoch = ++_reloadEpoch;
-    setState(() => _loading = true);
-    final result = await widget.notifier.loadNpcInventory(widget.npcId);
-    if (!mounted || epoch != _reloadEpoch) return;
-    setState(() {
-      _loading = false;
-      _result = result;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context);
-
-    if (_loading || _result == null) {
-      return const Padding(
-        padding: EdgeInsets.all(24),
-        child: Center(child: CircularProgressIndicator()),
-      );
-    }
-    final result = _result!;
-    if (result.error != null) {
-      return Padding(
-        padding: const EdgeInsets.all(20),
-        child: Text(
-          result.error!,
-          style: TextStyle(color: theme.colorScheme.error),
-        ),
-      );
-    }
-
-    final inventory = result.inventory;
-    final gates = _inventoryGates(
-      writable: inventory.writable,
-      privateEditable: widget.privateEditable,
-      typedVerified: widget.typedVerified,
-      canCompress: widget.canCompress,
-    );
-    final hasItems = inventory.hasData;
-    if (!hasItems && !gates.canAddItem && !gates.canRemoveItem) {
-      return _MessagePane(
-        icon: Icons.inventory_2_outlined,
-        title: l10n.inventoryTitle,
-        body: l10n.inventoryNoStacks,
-      );
-    }
-    return Padding(
-      padding: const EdgeInsets.all(20),
-      child: _PrivateInventorySummaryCard(
-        // Key on the NPC so switching NPCs builds a fresh card state (its
-        // local draft/pending fields reset) while the previous NPC's registry
-        // entry survives under its own per-NPC key.
-        key: ValueKey(('npc-inventory-card', widget.npcId)),
-        inventory: inventory,
-        notifier: widget.notifier,
-        editable: gates.editable,
-        canAddItem: gates.canAddItem,
-        canRemoveItem: gates.canRemoveItem,
-        pendingKey: 'inventory:${widget.npcId}',
-        actorId: widget.npcId,
-      ),
-    );
-  }
-}
-
-class _PrivateInventorySummaryCard extends ConsumerStatefulWidget {
-  const _PrivateInventorySummaryCard({
-    super.key,
-    required this.inventory,
-    required this.notifier,
-    this.editable = true,
-    this.canAddItem = false,
-    this.canRemoveItem = false,
-    this.pendingKey = 'inventory',
-    this.actorId,
-  });
-
-  final PrivateInventorySummary inventory;
-  final EditorNotifier notifier;
-  final bool editable;
-  final bool canAddItem;
-  final bool canRemoveItem;
-
-  /// Pending-edit registry key this card writes to. The player inventory uses
-  /// the literal `'inventory'`; each NPC uses a PER-NPC key (`'inventory:$id'`)
-  /// so editing NPC-A then NPC-B never clobbers either one's queued edits.
-  final String pendingKey;
-
-  /// GlobalId of the NPC whose inventory this edits, or null for the player.
-  /// Forwarded onto every queued edit as `actorId` so the core targets the
-  /// right container.
-  final String? actorId;
-
-  @override
-  ConsumerState<_PrivateInventorySummaryCard> createState() =>
-      _PrivateInventorySummaryCardState();
-}
-
-class _PrivateInventorySummaryCardState
-    extends ConsumerState<_PrivateInventorySummaryCard> {
-  String _query = '';
-  final TextEditingController _searchController = TextEditingController();
-  final Map<String, InventoryItemCountChange> _pendingCountChanges = {};
-  ItemCategory? _selectedCategory;
-  InventoryItemAdd? _pendingAdd;
-  // The item queued for removal (carries path + slotId + containerType so the
-  // core can target the right container's exact slot). addItem and removeItem
-  // are both structural edits, so at most one of _pendingAdd / _pendingRemove is
-  // ever set (the core allows one structural inventory edit per write).
-  PrivateInventoryItem? _pendingRemove;
-
-  /// Asset path of the item queued for removal, or null when none is queued.
-  /// Kept as a thin accessor so the existing path-based list filtering / display
-  /// is unchanged while [_pendingRemove] carries the full slot addressing.
-  String? get _pendingRemovePath => _pendingRemove?.path;
-
-  @override
-  void initState() {
-    super.initState();
-    // Rehydrate any draft already queued for THIS actor under the per-NPC
-    // pending key. Switching away from an edited NPC disposes this card's local
-    // state but leaves the registry entry (`inventory:<npcId>`) intact;
-    // revisiting builds a fresh card (keyed on the NPC) with empty local state.
-    // Without this, the next edit would call setPendingEdit with ONLY the new
-    // local edit and silently drop the earlier queued ones. Reading the stored
-    // edits back into local state makes a revisit resume from the queued draft.
-    _rehydrateFromPending();
-  }
-
-  /// Reconstruct local pending state from the registry entry for [pendingKey],
-  /// reversing the JSON edits [_pushInventoryPending] writes. Inverse of
-  /// [InventoryItemCountChange.toEditJson] et al. Tolerant of unexpected shapes
-  /// (skips anything it can't parse) so a malformed entry never throws here.
-  void _rehydrateFromPending() {
-    final entry = widget.notifier.pendingEditFor(widget.pendingKey);
-    if (entry == null) return;
-    for (final edit in entry.edits) {
-      final path = edit['path'];
-      final value = edit['value'];
-      if (value is! Map) continue;
-      switch (path) {
-        case 'private.inventory.setItemCount':
-          final itemPath = value['path'] as String? ?? '';
-          final count = (value['count'] as num?)?.toInt();
-          if (itemPath.isEmpty || count == null) continue;
-          final id = value['id'] as String? ?? '';
-          final slotId = (value['slotId'] as num?)?.toInt();
-          final containerType = value['containerType'] as String?;
-          final item = PrivateInventoryItem(
-            id: id,
-            path: itemPath,
-            slotId: slotId,
-            containerType: containerType,
-          );
-          _pendingCountChanges[_inventoryItemKey(item)] =
-              InventoryItemCountChange(
-            id: id,
-            path: itemPath,
-            count: count,
-            slotId: slotId,
-            containerType: containerType,
-          );
-        case 'private.inventory.addItem':
-          final itemPath = value['path'] as String? ?? '';
-          final count = (value['count'] as num?)?.toInt();
-          if (itemPath.isEmpty || count == null) continue;
-          _pendingAdd = InventoryItemAdd(path: itemPath, count: count);
-        case 'private.inventory.removeItem':
-          final itemPath = value['path'] as String? ?? '';
-          if (itemPath.isEmpty) continue;
-          _pendingRemove = PrivateInventoryItem(
-            id: value['id'] as String? ?? '',
-            path: itemPath,
-            slotId: (value['slotId'] as num?)?.toInt(),
-            containerType: value['containerType'] as String?,
-          );
-      }
-    }
-  }
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  @override
-  void didUpdateWidget(covariant _PrivateInventorySummaryCard oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.inventory != widget.inventory) {
-      // Inventory was refreshed — rebuild local widget state from the registry.
-      // A FULL save clears 'inventory:<id>' centrally, so rehydrate finds nothing
-      // and the drafts stay cleared. A PARTIALLY-committed save re-applies the
-      // still-uncommitted pending edits to the registry, so rehydrate restores
-      // them here — otherwise the Save badge would count queued edits while the
-      // count fields showed on-disk values, and the next edit would overwrite the
-      // registry with only that field. Rehydrate only mutates LOCAL state (never
-      // the provider), so it is safe in this build-context callback.
-      _pendingCountChanges.clear();
-      _pendingAdd = null;
-      _pendingRemove = null;
-      _rehydrateFromPending();
-    }
-  }
-
-  void _pushInventoryPending() {
-    final actorId = widget.actorId;
-    // Stamp the selected actor onto every queued edit. actorId is null for the
-    // player (key omitted from the JSON, behaviour byte-for-byte unchanged) and
-    // the NPC's GlobalId otherwise (core targets the NPC's container).
-    final countEdits = _pendingCountChanges.values
-        .map(
-          (c) => InventoryItemCountChange(
-            id: c.id,
-            path: c.path,
-            count: c.count,
-            actorId: actorId,
-            slotId: c.slotId,
-            containerType: c.containerType,
-          ).toEditJson(),
-        )
-        .toList();
-    final addEdit = _pendingAdd != null
-        ? InventoryItemAdd(
-            path: _pendingAdd!.path,
-            count: _pendingAdd!.count,
-            actorId: actorId,
-          ).toEditJson()
-        : null;
-    final removeEdit = _pendingRemove != null
-        ? InventoryItemRemove(
-            path: _pendingRemove!.path,
-            actorId: actorId,
-            slotId: _pendingRemove!.slotId,
-            containerType: _pendingRemove!.containerType,
-          ).toEditJson()
-        : null;
-    final allEdits = [
-      ...countEdits,
-      ?addEdit,
-      ?removeEdit,
-    ];
-    if (allEdits.isEmpty) {
-      widget.notifier.clearPendingEdit(widget.pendingKey);
-    } else {
-      widget.notifier.setPendingEdit(
-        widget.pendingKey,
-        PendingSaveEdit(edits: allEdits),
-      );
-    }
-  }
-
-  Future<void> _openAddDialog() async {
-    // Scope the dialog to the save it was opened for. If the user switches to a
-    // different save while the dialog is open, the awaited result is stale — its
-    // excludePaths and target belong to the old save, so applying it would queue
-    // the add against the wrong save. Key on the selected save path, not the
-    // inventory object identity: re-inspecting the SAME save allocates a fresh
-    // summary instance with identical contents, and that result must still apply.
-    final dialogSavePath = widget.notifier.selectedPath;
-    final result = await showDialog<InventoryItemAdd>(
-      context: context,
-      builder: (_) => AddInventoryItemDialog(
-        // Exclude both MainContainer paths (addItem rejects duplicates there)
-        // and the currently-equipped armor: adding a second copy of the worn
-        // armor would duplicate its definition path and make the equipped
-        // badge/upgrades ambiguous, so it must not be offered. Both lists are
-        // uncapped (from the typed tree), so the exclusion is correct even when
-        // the displayed row list is truncated.
-        excludePaths: {
-          ...widget.inventory.mainContainerPaths,
-          ...widget.inventory.equippedArmorPaths,
-        },
-      ),
-    );
-    if (result == null) return;
-    if (!mounted || widget.notifier.selectedPath != dialogSavePath) return;
-    setState(() {
-      _pendingAdd = result;
-      // Keep the one-structural-edit-per-save invariant unconditionally.
-      _pendingRemove = null;
-    });
-    _pushInventoryPending();
-  }
-
-  void _queueRemove(PrivateInventoryItem item) {
-    setState(() {
-      // A removal supersedes any pending count change on the same item, and is
-      // mutually exclusive with a pending add (one structural edit per save).
-      _pendingCountChanges.remove(_inventoryItemKey(item));
-      _pendingAdd = null;
-      // Carry the full item so the remove edit echoes slotId + containerType,
-      // letting the core target the exact slot in the right container.
-      _pendingRemove = item;
-    });
-    _pushInventoryPending();
-  }
-
-  void _undoRemove() {
-    setState(() => _pendingRemove = null);
-    _pushInventoryPending();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context);
-    final lang = ref.watch(currentGameLangProvider);
-    final locCatalog =
-        ref.watch(locCatalogProvider).asData?.value ?? const {};
-    final inventory = widget.inventory;
-    final query = _query.trim().toLowerCase();
-    final items = inventory.items.where((item) {
-      // A pending removal hides ONLY the specific slot queued for removal (it is
-      // represented by the pending card above), matched by the same slot-aware
-      // key as count edits — so duplicate-path stacks, or the same item in
-      // another container, are not all hidden when only one slot will be removed.
-      if (_pendingRemove != null &&
-          _inventoryItemKey(item) == _inventoryItemKey(_pendingRemove!)) {
-        return false;
-      }
-      if (query.isEmpty) return true;
-      final name = localizedGameName(locCatalog, lang, item.id);
-      return item.id.toLowerCase().contains(query) ||
-          item.path.toLowerCase().contains(query) ||
-          (name != null && name.toLowerCase().contains(query));
-    }).toList();
-    final groups = groupInventoryItems(items);
-
-    // Keep the current category selected if it still has items, else fall
-    // back to the first available group.
-    var selected = _selectedCategory;
-    if (groups.every((g) => g.category != selected)) {
-      selected = groups.isEmpty ? null : groups.first.category;
-    }
-    final selectedGroup =
-        groups.where((g) => g.category == selected).firstOrNull;
-
-    // An active search shows matches across all categories as a flat list;
-    // an empty query browses by the selected category.
-    final searching = query.isNotEmpty;
-    final shownItems = searching
-        ? items
-        : (selectedGroup?.items ?? const <PrivateInventoryItem>[]);
-
-    final hasItems = inventory.items.isNotEmpty;
-    final hasPendingAdd = _pendingAdd != null;
-    final hasPendingRemove = _pendingRemovePath != null;
-    final hasPendingCount = _pendingCountChanges.isNotEmpty;
-    final hasPendingChanges =
-        hasPendingCount || hasPendingAdd || hasPendingRemove;
-    final canRemove = widget.canRemoveItem;
-    // A structural edit (add/remove) must be saved on its own (the core/notifier
-    // reject a batch that mixes it with count edits), so structural edits and
-    // count edits are kept mutually exclusive in the UI: a structural edit is
-    // blocked while counts are pending, and count editing is blocked while a
-    // structural edit is pending.
-    final structuralBlocked = hasPendingAdd || hasPendingRemove || hasPendingCount;
-    final countEditable = widget.editable && !hasPendingAdd && !hasPendingRemove;
-
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                const Icon(Icons.inventory_2_outlined),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: Text(
-                    l10n.inventoryTitle,
-                    style: theme.textTheme.titleMedium,
-                  ),
-                ),
-                if (widget.editable && hasPendingChanges) ...[
-                  Tooltip(
-                    message: l10n.resetInventoryChanges,
-                    child: IconButton(
-                      icon: const Icon(Icons.undo_outlined),
-                      onPressed: () {
-                        setState(() {
-                          _pendingCountChanges.clear();
-                          _pendingAdd = null;
-                          _pendingRemove = null;
-                        });
-                        widget.notifier.clearPendingEdit(widget.pendingKey);
-                      },
-                    ),
-                  ),
-                ],
-                if (widget.canAddItem) ...[
-                  const SizedBox(width: 8),
-                  Tooltip(
-                    message: hasPendingAdd
-                        ? l10n.addItemTooltipPendingAdd
-                        : hasPendingRemove
-                            ? l10n.addItemTooltipPendingRemove
-                            : hasPendingCount
-                                ? l10n.addItemTooltipPendingCount
-                                : l10n.addItemTooltipDefault,
-                    child: FilledButton.icon(
-                      icon: const Icon(Icons.add, size: 18),
-                      label: Text(l10n.addItemButton),
-                      onPressed: structuralBlocked ? null : _openAddDialog,
-                    ),
-                  ),
-                ],
-              ],
-            ),
-            if (hasPendingAdd) ...[
-              const SizedBox(height: 12),
-              _PendingStructuralRow(
-                tone: _PendingTone.add,
-                icon: Icons.add_circle_outline,
-                title: _itemDisplayFromPath(_pendingAdd!.path),
-                subtitle: l10n.pendingAddSubtitle(_pendingAdd!.count),
-                cancelTooltip: l10n.cancelPendingAdd,
-                onCancel: () {
-                  setState(() => _pendingAdd = null);
-                  _pushInventoryPending();
-                },
-              ),
-            ],
-            if (hasPendingRemove) ...[
-              const SizedBox(height: 12),
-              _PendingStructuralRow(
-                tone: _PendingTone.remove,
-                icon: Icons.delete_outline,
-                title: _itemDisplayFromPath(_pendingRemovePath!),
-                subtitle: l10n.pendingRemovalSubtitle,
-                cancelTooltip: l10n.cancelPendingRemoval,
-                onCancel: _undoRemove,
-              ),
-            ],
-            if (hasItems) ...[
-              const SizedBox(height: 12),
-              TextField(
-                controller: _searchController,
-                decoration: InputDecoration(
-                  labelText: l10n.filterItems,
-                  prefixIcon: const Icon(Icons.search),
-                ),
-                onChanged: (value) => setState(() => _query = value),
-              ),
-              const SizedBox(height: 12),
-              Expanded(
-                child: groups.isEmpty
-                    ? Center(
-                        child: Text(
-                          // An empty query with no rows means a pending removal
-                          // hid the last item(s) — not a filter miss, so don't
-                          // claim "no items match".
-                          searching
-                              ? l10n.noItemsMatchQuery(_query)
-                              : l10n.pendingRemovalHidesAll,
-                          style: theme.textTheme.bodyMedium,
-                        ),
-                      )
-                    : Row(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          SizedBox(
-                            width: 200,
-                            child: DecoratedBox(
-                              decoration: BoxDecoration(
-                                color: theme.colorScheme.surfaceContainerLow,
-                                borderRadius: BorderRadius.circular(12),
-                              ),
-                              child: SingleChildScrollView(
-                                padding:
-                                    const EdgeInsets.symmetric(vertical: 6),
-                                child: Column(
-                                  children: [
-                                    for (final group in groups)
-                                      SidebarTile(
-                                        icon: iconForItemCategory(
-                                          group.category,
-                                        ),
-                                        label: l10n.categoryWithCount(
-                                          localizedItemCategoryLabel(
-                                            l10n,
-                                            group.category,
-                                          ),
-                                          group.items.length,
-                                        ),
-                                        selected: !searching &&
-                                            group.category == selected,
-                                        onTap: () => setState(() {
-                                          _selectedCategory = group.category;
-                                          // Leave search mode so the chosen
-                                          // category's items are shown.
-                                          _query = '';
-                                          _searchController.clear();
-                                        }),
-                                      ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                          ),
-                          const SizedBox(width: 16),
-                          Expanded(
-                            child: shownItems.isEmpty
-                                ? const SizedBox.shrink()
-                                : ListView.builder(
-                                    itemCount: shownItems.length,
-                                    itemBuilder: (context, index) {
-                                      final item = shownItems[index];
-                                      return ListTile(
-                                        dense: true,
-                                        leading: const Icon(
-                                          Icons.category_outlined,
-                                        ),
-                                        title: Row(
-                                          children: [
-                                            Flexible(
-                                              child: Text(
-                                                localizedGameName(
-                                                      locCatalog,
-                                                      lang,
-                                                      item.id,
-                                                    ) ??
-                                                    (item.id.isEmpty
-                                                        ? item.path
-                                                        : item.id),
-                                                maxLines: 1,
-                                                overflow: TextOverflow.ellipsis,
-                                              ),
-                                            ),
-                                            if (item.equipped) ...[
-                                              const SizedBox(width: 8),
-                                              Container(
-                                                padding:
-                                                    const EdgeInsets.symmetric(
-                                                      horizontal: 6,
-                                                      vertical: 2,
-                                                    ),
-                                                decoration: BoxDecoration(
-                                                  color: theme
-                                                      .colorScheme
-                                                      .primaryContainer,
-                                                  borderRadius:
-                                                      BorderRadius.circular(4),
-                                                ),
-                                                child: Text(
-                                                  l10n.equippedBadge,
-                                                  style: theme
-                                                      .textTheme
-                                                      .labelSmall
-                                                      ?.copyWith(
-                                                        color: theme
-                                                            .colorScheme
-                                                            .onPrimaryContainer,
-                                                      ),
-                                                ),
-                                              ),
-                                            ],
-                                          ],
-                                        ),
-                                        subtitle:
-                                            item.path.isEmpty &&
-                                                item.upgrades.isEmpty
-                                            ? null
-                                            : Column(
-                                                crossAxisAlignment:
-                                                    CrossAxisAlignment.start,
-                                                mainAxisSize: MainAxisSize.min,
-                                                children: [
-                                                  if (item.path.isNotEmpty)
-                                                    Text(
-                                                      item.path,
-                                                      maxLines: 1,
-                                                      overflow:
-                                                          TextOverflow.ellipsis,
-                                                    ),
-                                                  if (item.upgrades.isNotEmpty)
-                                                    Padding(
-                                                      padding:
-                                                          const EdgeInsets.only(
-                                                            top: 4,
-                                                          ),
-                                                      child: Wrap(
-                                                        spacing: 4,
-                                                        runSpacing: 2,
-                                                        crossAxisAlignment:
-                                                            WrapCrossAlignment
-                                                                .center,
-                                                        children: [
-                                                          Text(
-                                                            l10n.armorUpgradesLabel,
-                                                            style: theme
-                                                                .textTheme
-                                                                .labelSmall,
-                                                          ),
-                                                          for (final u
-                                                              in item.upgrades)
-                                                            Container(
-                                                              padding:
-                                                                  const EdgeInsets.symmetric(
-                                                                    horizontal: 6,
-                                                                    vertical: 1,
-                                                                  ),
-                                                              decoration: BoxDecoration(
-                                                                color: theme
-                                                                    .colorScheme
-                                                                    .surfaceContainerHighest,
-                                                                borderRadius:
-                                                                    BorderRadius.circular(
-                                                                      4,
-                                                                    ),
-                                                              ),
-                                                              child: Text(
-                                                                '${_upgradePart(u.key)}: ${_upgradeTier(u.value)}',
-                                                                style: theme
-                                                                    .textTheme
-                                                                    .labelSmall
-                                                                    ?.copyWith(
-                                                                      color: theme
-                                                                          .colorScheme
-                                                                          .onSurfaceVariant,
-                                                                    ),
-                                                              ),
-                                                            ),
-                                                        ],
-                                                      ),
-                                                    ),
-                                                ],
-                                              ),
-                                        trailing: _inventoryItemTrailing(
-                                          theme,
-                                          l10n,
-                                          item,
-                                          canRemove: canRemove,
-                                          countEditable: countEditable,
-                                          removeBlocked: structuralBlocked,
-                                        ),
-                                      );
-                                    },
-                                  ),
-                          ),
-                        ],
-                      ),
-              ),
-            ] else if (!hasPendingAdd) ...[
-              // Completely empty inventory (and nothing queued to add): say so
-              // explicitly instead of showing a blank card.
-              const SizedBox(height: 12),
-              Text(
-                l10n.inventoryEmpty,
-                style: theme.textTheme.bodyMedium?.copyWith(
-                  color: theme.colorScheme.onSurfaceVariant,
-                ),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _setPendingCountChange(
-    PrivateInventoryItem item,
-    InventoryItemCountChange? change,
-  ) {
-    setState(() {
-      final key = _inventoryItemKey(item);
-      if (change == null) {
-        _pendingCountChanges.remove(key);
-      } else {
-        _pendingCountChanges[key] = change;
-      }
-    });
-    _pushInventoryPending();
-  }
-
-  /// Trailing widget for an inventory row: count editor + a delete button.
-  Widget _inventoryItemTrailing(
-    ThemeData theme,
-    AppLocalizations l10n,
-    PrivateInventoryItem item, {
-    required bool canRemove,
-    required bool countEditable,
-    required bool removeBlocked,
-  }) {
-    // The count editor shows only when count editing is currently allowed (the
-    // inventory is count-editable AND no structural edit is pending). A
-    // remove-only inventory, or one with a pending structural edit, shows the
-    // count as plain text but the delete action may still apply.
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      // Centre the delete button against the count field so the trash icon lines
-      // up with the input value rather than floating up by the 'Count' label.
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        if (countEditable)
-          _InventoryItemCountEditor(
-            // Key by the slot-aware row key so duplicate-path/id stacks that
-            // differ only by slotId never share a controller (which could leave
-            // one row showing the other slot's count).
-            key: ValueKey(_inventoryItemKey(item)),
-            item: item,
-            pendingCount: _pendingCountChanges[_inventoryItemKey(item)]?.count,
-            onPendingCountChanged: (change) =>
-                _setPendingCountChange(item, change),
-          )
-        else
-          Text(
-            l10n.countTimes('${item.count ?? '?'}'),
-            style: theme.textTheme.bodyMedium,
-          ),
-        if (canRemove && item.path.isNotEmpty) ...[
-          const SizedBox(width: 4),
-          Tooltip(
-            message: !item.removable
-                ? l10n.deleteEquippedTooltip
-                : removeBlocked
-                ? l10n.removeBlockedTooltip
-                : l10n.removeItemFromInventory,
-            child: IconButton(
-              icon: const Icon(Icons.delete_outline),
-              // A non-removable item shows the trash icon disabled (its asset
-              // path occurs in more than one container — e.g. also equipped or
-              // in a quickslot — so the core can't unambiguously remove it). A
-              // removable item is disabled only while a structural/count edit
-              // is pending; otherwise it queues the remove.
-              onPressed: (!item.removable || removeBlocked)
-                  ? null
-                  : () => _queueRemove(item),
-            ),
-          ),
-        ],
-      ],
-    );
-  }
-}
-
-/// Tone of a pending structural-edit card: an add (primary) or a remove
-/// (error).
-enum _PendingTone { add, remove }
-
-/// A human-readable id fragment derived from an item asset path.
-String _itemDisplayFromPath(String path) =>
-    path.contains('.') ? path.split('.').last : path.split('/').last;
-
-String _upgradePart(String key) {
-  if (key.contains('Upper')) return 'Upper';
-  if (key.contains('Mid')) return 'Mid';
-  if (key.contains('Lower')) return 'Lower';
-  return key;
-}
-
-String _upgradeTier(String value) {
-  var v = value;
-  for (final p in const ['m_UpperBody_', 'm_MidBody_', 'm_LowerBody_']) {
-    if (v.startsWith(p)) {
-      v = v.substring(p.length);
-      break;
-    }
-  }
-  return v.replaceAll('_ArmorUpgrade', '');
-}
-
-/// A highlighted card shown when there is a pending structural inventory edit
-/// (add or remove) awaiting save. Mirrors how a not-yet-saved item is
-/// represented for both directions: the affected item is not shown inline, only
-/// here, with a cancel button.
-class _PendingStructuralRow extends StatelessWidget {
-  const _PendingStructuralRow({
-    required this.icon,
-    required this.title,
-    required this.subtitle,
-    required this.onCancel,
-    required this.cancelTooltip,
-    required this.tone,
-  });
-
-  final IconData icon;
-  final String title;
-  final String subtitle;
-  final VoidCallback onCancel;
-  final String cancelTooltip;
-  final _PendingTone tone;
-
-  @override
-  Widget build(BuildContext context) {
-    final scheme = Theme.of(context).colorScheme;
-    final isAdd = tone == _PendingTone.add;
-    final bg = isAdd ? scheme.primaryContainer : scheme.errorContainer;
-    final fg = isAdd ? scheme.onPrimaryContainer : scheme.onErrorContainer;
-    final accent = isAdd ? scheme.primary : scheme.error;
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: bg,
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: accent.withValues(alpha: 0.4)),
-      ),
-      child: ListTile(
-        dense: true,
-        leading: Icon(icon, color: accent),
-        title: Text(
-          title,
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-          style: TextStyle(color: fg),
-        ),
-        subtitle: Text(
-          subtitle,
-          style: TextStyle(color: fg.withValues(alpha: 0.8)),
-        ),
-        trailing: IconButton(
-          icon: const Icon(Icons.close),
-          tooltip: cancelTooltip,
-          onPressed: onCancel,
-        ),
-      ),
-    );
-  }
-}
-
-class _InventoryItemCountEditor extends StatefulWidget {
-  const _InventoryItemCountEditor({
-    super.key,
-    required this.item,
-    required this.onPendingCountChanged,
-    this.pendingCount,
-  });
-
-  final PrivateInventoryItem item;
-  final int? pendingCount;
-  final void Function(InventoryItemCountChange? change) onPendingCountChanged;
-
-  @override
-  State<_InventoryItemCountEditor> createState() =>
-      _InventoryItemCountEditorState();
-}
-
-class _InventoryItemCountEditorState extends State<_InventoryItemCountEditor> {
-  late final TextEditingController _controller;
-  String? _path;
-  String? _id;
-  int? _slotId;
-  int? _count;
-  int? _pendingCount;
-  String? _error;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = TextEditingController();
-    _sync();
-  }
-
-  @override
-  void didUpdateWidget(covariant _InventoryItemCountEditor oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    _sync();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _sync() {
-    // Rows are identified by path-or-id PLUS slotId — two duplicate stacks share
-    // a path/id and differ only by slotId, and the on-disk count can change under
-    // a fixed slot after a refresh — so include slotId and the canonical count;
-    // otherwise the early-return could leave the field showing a stale count.
-    if (_path == widget.item.path &&
-        _id == widget.item.id &&
-        _slotId == widget.item.slotId &&
-        _count == widget.item.count &&
-        _pendingCount == widget.pendingCount) {
-      return;
-    }
-    final isSameItem = _path == widget.item.path &&
-        _id == widget.item.id &&
-        _slotId == widget.item.slotId;
-    _path = widget.item.path;
-    _id = widget.item.id;
-    _slotId = widget.item.slotId;
-    _count = widget.item.count;
-    _pendingCount = widget.pendingCount;
-    final text = (widget.pendingCount ?? widget.item.count)?.toString() ?? '';
-    if (_controller.text != text) {
-      final currentOffset = _controller.selection.baseOffset;
-      final nextOffset = isSameItem
-          ? currentOffset.clamp(0, text.length)
-          : text.length;
-      _controller.value = TextEditingValue(
-        text: text,
-        selection: TextSelection.collapsed(offset: nextOffset),
-      );
-    }
-    _error = null;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 132,
-      child: TextField(
-        controller: _controller,
-        keyboardType: TextInputType.number,
-        onChanged: _onCountTextChanged,
-        decoration: InputDecoration(
-          labelText: AppLocalizations.of(context).count,
-          errorText: _error,
-          // Compact the field so it fits inside the dense ListTile row. A
-          // reserved helper line would steal the input box's vertical space
-          // here (the tile caps the field height), squeezing the box until the
-          // value clips at the border — so the error grows the row instead.
-          isDense: true,
-        ),
-      ),
-    );
-  }
-
-  void _onCountTextChanged(String value) {
-    final trimmed = value.trim();
-    if (trimmed.isEmpty) {
-      setState(() => _error = null);
-      widget.onPendingCountChanged(null);
-      return;
-    }
-    final parsed = int.tryParse(trimmed);
-    if (parsed == null || parsed < 1) {
-      // Min 1: a count of 0 would leave a ghost slot (invisible in-game but
-      // still in the save). Use the remove button to delete an item.
-      final min1 = AppLocalizations.of(context).min1;
-      setState(() => _error = min1);
-      widget.onPendingCountChanged(null);
-      return;
-    }
-    setState(() => _error = null);
-    if (parsed == widget.item.count) {
-      widget.onPendingCountChanged(null);
-      return;
-    }
-    widget.onPendingCountChanged(
-      InventoryItemCountChange(
-        id: widget.item.id,
-        path: widget.item.path,
-        count: parsed,
-        slotId: widget.item.slotId,
-        containerType: widget.item.containerType,
-      ),
-    );
-  }
-}
-
-String _inventoryItemKey(PrivateInventoryItem item) {
-  // Combine container type, slot id, id, and path so rows that share a
-  // definition path get distinct pending-change entries instead of collapsing
-  // onto one key. The container type keeps identical paths in different
-  // containers (e.g. a MeleeSlot sword and a Pouch stack sharing a slotId)
-  // distinct; the stable slot id keeps two duplicate same-path/same-id stacks in
-  // ONE container independent; id+path remains the fallback for rows without one.
-  return '${item.containerType ?? ''}\u0000${item.slotId ?? ''}\u0000${item.id}\u0000${item.path}';
-}
-
-class _PrivatePlayerTransformEditor extends StatefulWidget {
-  const _PrivatePlayerTransformEditor({
-    required this.transform,
-    required this.editable,
-    required this.notifier,
-    this.reloadKey,
-  });
-
-  final PrivatePlayerTransform transform;
-  final bool editable;
-  final EditorNotifier notifier;
-  // When provided, a change in identity triggers a field reseed even if the
-  // transform values haven't changed (e.g. Reset followed by re-inspect that
-  // returns the same values).
-  final Object? reloadKey;
-
-  @override
-  State<_PrivatePlayerTransformEditor> createState() =>
-      _PrivatePlayerTransformEditorState();
-}
-
-class _PrivatePlayerTransformEditorState
-    extends State<_PrivatePlayerTransformEditor> {
-  late final TextEditingController _locationXController;
-  late final TextEditingController _locationYController;
-  late final TextEditingController _locationZController;
-  late final TextEditingController _rotationPitchController;
-  late final TextEditingController _rotationYawController;
-  late final TextEditingController _rotationRollController;
-  PrivatePlayerTransform? _lastTransform;
-  // Track the inspection (widget parent) identity so that a Reset/refresh that
-  // produces a new inspection instance triggers a reseed even when the
-  // transform values themselves haven't changed.
-  Object? _inspectionIdentity;
-  String? _error;
-
-  @override
-  void initState() {
-    super.initState();
-    _locationXController = TextEditingController();
-    _locationYController = TextEditingController();
-    _locationZController = TextEditingController();
-    _rotationPitchController = TextEditingController();
-    _rotationYawController = TextEditingController();
-    _rotationRollController = TextEditingController();
-    _sync();
-  }
-
-  @override
-  void didUpdateWidget(covariant _PrivatePlayerTransformEditor oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    _sync();
-  }
-
-  @override
-  void dispose() {
-    _locationXController.dispose();
-    _locationYController.dispose();
-    _locationZController.dispose();
-    _rotationPitchController.dispose();
-    _rotationYawController.dispose();
-    _rotationRollController.dispose();
-    super.dispose();
-  }
-
-  void _sync() {
-    final transform = widget.transform;
-    final last = _lastTransform;
-    // Re-seed on reloadKey identity change (e.g. after Reset/refresh that
-    // produces a new SaveInspection) even when the transform values themselves
-    // are unchanged, so the fields visually revert after a Reset.
-    final newKey = widget.reloadKey;
-    final sameKey = newKey == null || identical(newKey, _inspectionIdentity);
-    if (!sameKey) {
-      _inspectionIdentity = newKey;
-    }
-    if (sameKey &&
-        last != null &&
-        last.location.x == transform.location.x &&
-        last.location.y == transform.location.y &&
-        last.location.z == transform.location.z &&
-        last.rotation.pitch == transform.rotation.pitch &&
-        last.rotation.yaw == transform.rotation.yaw &&
-        last.rotation.roll == transform.rotation.roll) {
-      return;
-    }
-    _lastTransform = transform;
-    _locationXController.text = _formatAttributeValue(transform.location.x);
-    _locationYController.text = _formatAttributeValue(transform.location.y);
-    _locationZController.text = _formatAttributeValue(transform.location.z);
-    _rotationPitchController.text = _formatAttributeValue(
-      transform.rotation.pitch,
-    );
-    _rotationYawController.text = _formatAttributeValue(transform.rotation.yaw);
-    _rotationRollController.text = _formatAttributeValue(
-      transform.rotation.roll,
-    );
-    _error = null;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            const Icon(Icons.explore_outlined),
-            const SizedBox(width: 8),
-            Expanded(
-              child: Text(
-                l10n.heroTransform,
-                style: Theme.of(context).textTheme.titleSmall,
-              ),
-            ),
-          ],
-        ),
-        if (_error != null) ...[
-          const SizedBox(height: 6),
-          Text(
-            _error!,
-            style: TextStyle(color: Theme.of(context).colorScheme.error),
-          ),
-        ],
-        const SizedBox(height: 10),
-        LayoutBuilder(
-          builder: (context, constraints) {
-            final compact = constraints.maxWidth < 700;
-            final fields = [
-              _TransformNumberField(
-                controller: _locationXController,
-                label: l10n.locationX,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-              _TransformNumberField(
-                controller: _locationYController,
-                label: l10n.locationY,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-              _TransformNumberField(
-                controller: _locationZController,
-                label: l10n.locationZ,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-              _TransformNumberField(
-                controller: _rotationPitchController,
-                label: l10n.rotationPitch,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-              _TransformNumberField(
-                controller: _rotationYawController,
-                label: l10n.rotationYaw,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-              _TransformNumberField(
-                controller: _rotationRollController,
-                label: l10n.rotationRoll,
-                enabled: widget.editable,
-                onChanged: (_) => _updatePending(),
-              ),
-            ];
-            if (compact) {
-              return Column(
-                children: [
-                  for (final field in fields) ...[
-                    field,
-                    if (field != fields.last) const SizedBox(height: 8),
-                  ],
-                ],
-              );
-            }
-            return Column(
-              children: [
-                Row(
-                  children: [
-                    for (final field in fields.take(3)) ...[
-                      Expanded(child: field),
-                      if (field != fields[2]) const SizedBox(width: 8),
-                    ],
-                  ],
-                ),
-                const SizedBox(height: 8),
-                Row(
-                  children: [
-                    for (final field in fields.skip(3)) ...[
-                      Expanded(child: field),
-                      if (field != fields.last) const SizedBox(width: 8),
-                    ],
-                  ],
-                ),
-              ],
-            );
-          },
-        ),
-      ],
-    );
-  }
-
-  void _updatePending() {
-    if (!widget.editable) return;
-    final locationX = double.tryParse(_locationXController.text.trim());
-    final locationY = double.tryParse(_locationYController.text.trim());
-    final locationZ = double.tryParse(_locationZController.text.trim());
-    final rotationPitch = double.tryParse(_rotationPitchController.text.trim());
-    final rotationYaw = double.tryParse(_rotationYawController.text.trim());
-    final rotationRoll = double.tryParse(_rotationRollController.text.trim());
-    if (locationX == null ||
-        locationY == null ||
-        locationZ == null ||
-        rotationPitch == null ||
-        rotationYaw == null ||
-        rotationRoll == null) {
-      final invalid = AppLocalizations.of(context).invalid;
-      setState(() => _error = invalid);
-      widget.notifier.clearPendingEdit('transform');
-      return;
-    }
-    setState(() => _error = null);
-    final orig = widget.transform;
-    if (locationX == orig.location.x &&
-        locationY == orig.location.y &&
-        locationZ == orig.location.z &&
-        rotationPitch == orig.rotation.pitch &&
-        rotationYaw == orig.rotation.yaw &&
-        rotationRoll == orig.rotation.roll) {
-      widget.notifier.clearPendingEdit('transform');
-      return;
-    }
-    widget.notifier.setPendingEdit(
-      'transform',
-      PendingSaveEdit(
-        edits: [
-          {
-            'path': 'private.player.setTransform',
-            'value': {
-              'location': {'x': locationX, 'y': locationY, 'z': locationZ},
-              'rotation': {
-                'pitch': rotationPitch,
-                'yaw': rotationYaw,
-                'roll': rotationRoll,
-              },
-            },
-          },
-        ],
-      ),
-    );
-  }
-}
-
-class _TransformNumberField extends StatelessWidget {
-  const _TransformNumberField({
-    required this.controller,
-    required this.label,
-    required this.enabled,
-    this.onChanged,
-  });
-
-  final TextEditingController controller;
-  final String label;
-  final bool enabled;
-  final ValueChanged<String>? onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      controller: controller,
-      enabled: enabled,
-      onChanged: onChanged,
-      keyboardType: const TextInputType.numberWithOptions(
-        decimal: true,
-        signed: true,
-      ),
-      decoration: InputDecoration(labelText: label),
-    );
-  }
-}
-
-class _PrivatePlayerAttributesEditor extends StatelessWidget {
-  const _PrivatePlayerAttributesEditor({
-    required this.player,
-    required this.notifier,
-    this.editable = true,
-    this.reloadKey,
-  });
-
-  final PrivatePlayerSummary player;
-  final EditorNotifier notifier;
-  final bool editable;
-  final Object? reloadKey;
-
-  @override
-  Widget build(BuildContext context) {
-    final editable =
-        this.editable &&
-        player.writable.contains('private.player.setAttribute');
-    final l10n = AppLocalizations.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            const Icon(Icons.monitor_heart_outlined),
-            const SizedBox(width: 8),
-            Text(
-              l10n.heroAttributes,
-              style: Theme.of(context).textTheme.titleSmall,
-            ),
-          ],
-        ),
-        const SizedBox(height: 10),
-        LayoutBuilder(
-          builder: (context, constraints) {
-            final compact = constraints.maxWidth < 620;
-            return Column(
-              children: player.attributes
-                  .map(
-                    (attribute) => _PrivatePlayerAttributeRow(
-                      attribute: attribute,
-                      notifier: notifier,
-                      editable: editable,
-                      compact: compact,
-                      reloadKey: reloadKey,
-                    ),
-                  )
-                  .toList(),
-            );
-          },
-        ),
-      ],
-    );
-  }
-}
-
-class _PrivatePlayerAttributeRow extends StatefulWidget {
-  const _PrivatePlayerAttributeRow({
-    required this.attribute,
-    required this.notifier,
-    required this.editable,
-    required this.compact,
-    this.reloadKey,
-  });
-
-  final PrivatePlayerAttribute attribute;
-  final EditorNotifier notifier;
-  final bool editable;
-  final bool compact;
-  // When provided, a change in identity triggers a field reseed even if the
-  // attribute values haven't changed (e.g. after a Reset that reverts to the
-  // same canonical value).
-  final Object? reloadKey;
-
-  @override
-  State<_PrivatePlayerAttributeRow> createState() =>
-      _PrivatePlayerAttributeRowState();
-}
-
-class _PrivatePlayerAttributeRowState
-    extends State<_PrivatePlayerAttributeRow> {
-  late final TextEditingController _baseController;
-  late final TextEditingController _currentController;
-  String? _lastId;
-  double? _lastBase;
-  double? _lastCurrent;
-  Object? _lastReloadKey;
-  String? _error;
-
-  @override
-  void initState() {
-    super.initState();
-    _baseController = TextEditingController();
-    _currentController = TextEditingController();
-    _sync();
-  }
-
-  @override
-  void didUpdateWidget(covariant _PrivatePlayerAttributeRow oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    _sync();
-  }
-
-  @override
-  void dispose() {
-    _baseController.dispose();
-    _currentController.dispose();
-    super.dispose();
-  }
-
-  void _sync() {
-    final attribute = widget.attribute;
-    final newKey = widget.reloadKey;
-    final sameKey = newKey == null || identical(newKey, _lastReloadKey);
-    if (!sameKey) {
-      _lastReloadKey = newKey;
-    }
-    if (sameKey &&
-        _lastId == attribute.id &&
-        _lastBase == attribute.baseValue &&
-        _lastCurrent == attribute.currentValue) {
-      return;
-    }
-    _lastId = attribute.id;
-    _lastBase = attribute.baseValue;
-    _lastCurrent = attribute.currentValue;
-    _baseController.text = _formatAttributeValue(attribute.baseValue);
-    _currentController.text = _formatAttributeValue(attribute.currentValue);
-    _error = null;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context);
-    final name = widget.attribute.id;
-    final baseField = TextField(
-      controller: _baseController,
-      enabled: widget.editable,
-      keyboardType: const TextInputType.numberWithOptions(decimal: true),
-      onChanged: (_) => _updatePending(),
-      decoration: InputDecoration(
-        labelText: l10n.attributeBase(name),
-        errorText: _error,
-      ),
-    );
-    final currentField = TextField(
-      controller: _currentController,
-      enabled: widget.editable,
-      keyboardType: const TextInputType.numberWithOptions(decimal: true),
-      onChanged: (_) => _updatePending(),
-      decoration: InputDecoration(labelText: l10n.attributeCurrent(name)),
-    );
-    final label = SizedBox(
-      width: 116,
-      child: Text(name, style: Theme.of(context).textTheme.labelLarge),
-    );
-    if (widget.compact) {
-      return Padding(
-        padding: const EdgeInsets.symmetric(vertical: 6),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Text(name, style: Theme.of(context).textTheme.labelLarge),
-            const SizedBox(height: 6),
-            baseField,
-            const SizedBox(height: 6),
-            currentField,
-          ],
-        ),
-      );
-    }
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 5),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          label,
-          Expanded(child: baseField),
-          const SizedBox(width: 8),
-          Expanded(child: currentField),
-        ],
-      ),
-    );
-  }
-
-  void _updatePending() {
-    if (!widget.editable) return;
-    final id = widget.attribute.id;
-    final baseValue = double.tryParse(_baseController.text.trim());
-    final currentValue = double.tryParse(_currentController.text.trim());
-    if (baseValue == null || currentValue == null) {
-      final invalid = AppLocalizations.of(context).invalid;
-      setState(() => _error = invalid);
-      widget.notifier.clearPendingEdit('attr:$id');
-      return;
-    }
-    setState(() => _error = null);
-    final origBase = widget.attribute.baseValue;
-    final origCurrent = widget.attribute.currentValue;
-    if (baseValue == origBase && currentValue == origCurrent) {
-      widget.notifier.clearPendingEdit('attr:$id');
-      return;
-    }
-    widget.notifier.setPendingEdit(
-      'attr:$id',
-      PendingSaveEdit(
-        edits: [
-          {
-            'path': 'private.player.setAttribute',
-            'value': {
-              'id': id,
-              'baseValue': baseValue,
-              'currentValue': currentValue,
-            },
-          },
-        ],
-      ),
-    );
-  }
-}
-
-String _formatAttributeValue(double? value) {
-  if (value == null) return '';
-  if (value == value.roundToDouble()) return value.toInt().toString();
-  final rounded = value.toStringAsFixed(2).replaceFirst(RegExp(r'\.?0+$'), '');
-  // These texts seed editable fields whose parsed value gets written back —
-  // a lossy rounding (0.125 → 0.13) would silently corrupt untouched axes
-  // the moment any sibling field changes. Round-trip or full precision.
-  return double.tryParse(rounded) == value ? rounded : value.toString();
 }
 
 /// Generic typed property browser: search every property in the decoded
