+++ apps/save-editor/lib/l10n/app_en.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "Attributes",
   "tabInventory": "Inventory",
   "tabProgression": "Progression",
+  "tabCharacters": "Characters",
+  "characterNoActorBody": "This character has no in-world actor, so it has no attributes, inventory, or events.",
   "tabAllData": "All data",
   "tabBackups": "Backups",
   "tabSettings": "Settings",
