+++ apps/save-editor/lib/l10n/app_es.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "Atributos",
   "tabInventory": "Inventario",
   "tabProgression": "Progreso",
+  "tabCharacters": "Personajes",
+  "characterNoActorBody": "Este personaje no tiene un actor en el mundo, por lo que no tiene atributos, inventario ni eventos.",
   "tabAllData": "Todos los datos",
   "tabBackups": "Copias de seguridad",
   "tabSettings": "Ajustes",
