+++ apps/save-editor/lib/l10n/app_fr.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "Attributs",
   "tabInventory": "Inventaire",
   "tabProgression": "Progression",
+  "tabCharacters": "Personnages",
+  "characterNoActorBody": "Ce personnage n'a pas d'acteur dans le monde ; il n'a donc ni attributs, ni inventaire, ni événements.",
   "tabAllData": "Toutes les données",
   "tabBackups": "Sauvegardes",
   "tabSettings": "Paramètres",
