+++ apps/save-editor/lib/l10n/app_it.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "Attributi",
   "tabInventory": "Inventario",
   "tabProgression": "Progressione",
+  "tabCharacters": "Personaggi",
+  "characterNoActorBody": "Questo personaggio non ha un attore nel mondo, quindi non ha attributi, inventario o eventi.",
   "tabAllData": "Tutti i dati",
   "tabBackups": "Backup",
   "tabSettings": "Impostazioni",
