+++ apps/save-editor/lib/l10n/app_pt_BR.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "Atributos",
   "tabInventory": "Inventário",
   "tabProgression": "Progresso",
+  "tabCharacters": "Personagens",
+  "characterNoActorBody": "Este personagem não tem um ator no mundo, portanto não tem atributos, inventário ou eventos.",
   "tabAllData": "Todos os dados",
   "tabBackups": "Backups",
   "tabSettings": "Configurações",
