+++ apps/save-editor/lib/l10n/app_zh.arb
@@ -11,6 +11,8 @@
   "tabAttribute": "属性",
   "tabInventory": "物品栏",
   "tabProgression": "进度",
+  "tabCharacters": "角色",
+  "characterNoActorBody": "该角色在世界中没有对应的实体，因此没有属性、物品栏或事件。",
   "tabAllData": "全部数据",
   "tabBackups": "备份",
   "tabSettings": "设置",
