+++ apps/save-editor/test/npc_inventory_duplicate_stack_count_test.dart
@@ -41,6 +41,8 @@ void main() {
       final core = _DuplicateStackNpcInventoryCoreService();
       await pumpApp(tester, core);
 
+      await tester.tap(find.widgetWithText(Tab, 'Characters'));
+      await tester.pumpAndSettle();
       await tester.tap(find.widgetWithText(Tab, 'Inventory'));
       await tester.pumpAndSettle();
       await tester.tap(find.text('Lizard-A'));
@@ -78,6 +80,8 @@ void main() {
       final core = _DuplicateStackNpcInventoryCoreService();
       await pumpApp(tester, core);
 
+      await tester.tap(find.widgetWithText(Tab, 'Characters'));
+      await tester.pumpAndSettle();
       await tester.tap(find.widgetWithText(Tab, 'Inventory'));
       await tester.pumpAndSettle();
       await tester.tap(find.text('Lizard-A'));
@@ -203,6 +207,34 @@ class _DuplicateStackNpcInventoryCoreService implements GoresaveCoreService {
             'status': 'ready',
             'adapter': 'pure_rust_kraken',
             'message': 'Codec host is ready.',
+          },
+        };
+      case 'private.characters.list':
+        // Backs the Charaktere master list (one unpaginated response). The
+        // globalId is rendered as the row subtitle, so tests select a row by
+        // tapping its GlobalId text (e.g. 'Lizard-A').
+        return {
+          'ok': true,
+          'data': {
+            'total': 2,
+            'characters': [
+              {
+                'globalId': 'Lizard-A',
+                'uniqueName': 'Lizard',
+                'isDead': false,
+                'hasInventory': false,
+                'hasKnowledge': false,
+                'hasEvents': false,
+              },
+              {
+                'globalId': 'Lizard-B',
+                'uniqueName': 'Lizard',
+                'isDead': false,
+                'hasInventory': false,
+                'hasKnowledge': false,
+                'hasEvents': false,
+              },
+            ],
           },
         };
       case 'private.npc.list':
