+++ apps/save-editor/test/npc_inventory_pending_per_npc_test.dart
@@ -35,7 +35,9 @@ void main() {
       final core = _NpcInventoryCoreService();
       await pumpApp(tester, core);
 
-      // Open the Inventory tab (hosts the shared ActorSelector + inventory card).
+      // Open the Charaktere tab (shared master list) then its Inventar sub-tab.
+      await tester.tap(find.widgetWithText(Tab, 'Characters'));
+      await tester.pumpAndSettle();
       await tester.tap(find.widgetWithText(Tab, 'Inventory'));
       await tester.pumpAndSettle();
 
@@ -96,13 +98,16 @@ void main() {
       final core = _NpcInventoryCoreService();
       await pumpApp(tester, core);
 
+      await tester.tap(find.widgetWithText(Tab, 'Characters'));
+      await tester.pumpAndSettle();
       await tester.tap(find.widgetWithText(Tab, 'Inventory'));
       await tester.pumpAndSettle();
       await tester.tap(find.text('Lizard-B'));
       await tester.pumpAndSettle();
 
-      // The Attribute tab reads the SAME selectedActor — its NPC attributes
-      // panel loads Lizard-B (the fake returns Lizard-B's attribute row).
+      // The Attribute sub-tab reads the SAME shared selectedActor — switching to
+      // it (no re-select) loads Lizard-B's NPC attributes (the fake returns
+      // Lizard-B's attribute row), proving selection is shared across sub-tabs.
       await tester.tap(find.widgetWithText(Tab, 'Attributes'));
       await tester.pumpAndSettle();
       final attrReq = core.requests.lastWhere(
@@ -224,6 +229,34 @@ class _NpcInventoryCoreService implements GoresaveCoreService {
             'status': 'ready',
             'adapter': 'pure_rust_kraken',
             'message': 'Codec host is ready.',
+          },
+        };
+      case 'private.characters.list':
+        // Backs the Charaktere master list (one unpaginated response). The
+        // globalId is rendered as the row subtitle, so tests select a row by
+        // tapping its GlobalId text (e.g. 'Lizard-A').
+        return {
+          'ok': true,
+          'data': {
+            'total': 2,
+            'characters': [
+              {
+                'globalId': 'Lizard-A',
+                'uniqueName': 'Lizard',
+                'isDead': false,
+                'hasInventory': false,
+                'hasKnowledge': false,
+                'hasEvents': false,
+              },
+              {
+                'globalId': 'Lizard-B',
+                'uniqueName': 'Lizard',
+                'isDead': false,
+                'hasInventory': false,
+                'hasKnowledge': false,
+                'hasEvents': false,
+              },
+            ],
           },
         };
       case 'private.npc.list':
